[Feature] Move all resize script into resizevolume dir
